refactor(repository): restrict repositories visibility
