+++ app/src/main/java/com/example/shmr_finance_app_android/core/di/RepositoryModule.kt
@@ -14,7 +14,7 @@ import javax.inject.Singleton
 
 @Module
 @InstallIn(SingletonComponent::class)
-interface RepositoryModule {
+internal interface RepositoryModule {
 
     @Binds
     @Singleton
