+++ app/src/main/java/com/example/shmr_finance_app_android/data/repository/TransactionsRepositoryImpl.kt
@@ -4,12 +4,12 @@ import android.os.Build
 import androidx.annotation.RequiresApi
 import com.example.shmr_finance_app_android.data.datasource.TransactionsRemoteDataSource
 import com.example.shmr_finance_app_android.data.remote.api.safeApiCall
-import com.example.shmr_finance_app_android.domain.mapper.TransactionsDomainMapper
+import com.example.shmr_finance_app_android.data.repository.mapper.TransactionsDomainMapper
 import com.example.shmr_finance_app_android.domain.model.TransactionDomain
 import com.example.shmr_finance_app_android.domain.repository.TransactionsRepository
 import javax.inject.Inject
 
-class TransactionsRepositoryImpl @Inject constructor(
+internal class TransactionsRepositoryImpl @Inject constructor(
     private val remoteDataSource: TransactionsRemoteDataSource,
     private val mapper: TransactionsDomainMapper
 ) : TransactionsRepository {
